+++ src/Pane.ts
@@ -33,6 +33,13 @@ namespace g {
 		 * @default undefined
 		 */
 		backgroundEffector?: SurfaceEffector;
+
+		/**
+		 * この`Pane`をスケール変更に対応させるか指定する
+		 * falseもしくはundefinedの場合、スケール変更に対応させない
+		 * @default undefined
+		 */
+		hasVariableResolution?: boolean;
 	}
 	/**
 	 * 枠を表すエンティティ。
@@ -107,6 +114,11 @@ namespace g {
 		_childrenRenderer: Renderer;
 
 		/**
+		 * @private
+		 */
+		_hasVariableResolution: boolean;
+
+		/**
 		 * 各種パラメータを指定して `Pane` のインスタンスを生成する。
 		 * @param param このエンティティに指定するパラメータ
 		 */
@@ -118,6 +130,7 @@ namespace g {
 			this.backgroundEffector = param.backgroundEffector;
 			this._shouldRenderChildren = false;
 			this._padding = param.padding;
+			this._hasVariableResolution = param.hasVariableResolution !== undefined ? param.hasVariableResolution : false;
 			this._initialize();
 			this._paddingChanged = false;
 			this._bgSurface = undefined;
@@ -263,7 +276,12 @@ namespace g {
 			if (this._childrenSurface && !this._childrenSurface.destroyed()) {
 				this._childrenSurface.destroy();
 			}
-			this._childrenSurface = resourceFactory.createSurface(Math.ceil(this._childrenArea.width), Math.ceil(this._childrenArea.height));
+			const surfaceStateFlag = this._hasVariableResolution ? g.SurfaceStateFlags.hasVariableResolution : g.SurfaceStateFlags.None;
+			this._childrenSurface = resourceFactory.createSurface(
+				Math.ceil(this._childrenArea.width),
+				Math.ceil(this._childrenArea.height),
+				surfaceStateFlag
+			);
 			this._childrenRenderer = this._childrenSurface.renderer();
 			this._normalizedPadding = r;
 		}
@@ -306,6 +324,10 @@ namespace g {
 					result.bottom = convertedPoint.y;
 			}
 			return result;
+		}
+
+		_createCache(width: number, height: number): Surface {
+			return this._createSurface(width, height, this._hasVariableResolution);
 		}
 	}
 }